backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
apscheduler>=3.10.4
pandas>=2.2.0
numpy>=2.0.0
xgboost>=2.1.0
scikit-learn>=1.5.0
joblib>=1.4.0
requests>=2.32.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
google-generativeai>=0.8.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
